Convert popup to animated dialog, add delete option
Converted the popup to an animated dialof with a background picture of
a bookshelf. Added a deleted option on the end of each row, that deletes
the book information from the row and removes the book data from the
myLibrary array. Could add a clear all option for the displayed table,
or prevent table from adding more content if all are displayed.